+++ noodles-vcf/src/header/record/value/map/tag/other.rs
@@ -1,8 +1,9 @@
 use std::{
     borrow::Borrow,
-    fmt,
+    error, fmt,
     hash::{Hash, Hasher},
     marker::PhantomData,
+    str::FromStr,
 };
 
 use super::Standard;
@@ -45,5 +46,54 @@ impl<S> Eq for Other<S> {}
 impl<S> fmt::Display for Other<S> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         self.0.fmt(f)
+    }
+}
+
+/// An error returned when a raw VCF header record value map other tag fails to a parse.
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub enum ParseError {
+    /// The input is invalid.
+    Invalid,
+}
+
+impl error::Error for ParseError {}
+
+impl fmt::Display for ParseError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Invalid => write!(f, "invalid input"),
+        }
+    }
+}
+
+impl<S> FromStr for Other<S>
+where
+    S: Standard,
+{
+    type Err = ParseError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        use super::Tag;
+
+        match Tag::from(s.to_string()) {
+            Tag::Standard(_) => Err(ParseError::Invalid),
+            Tag::Other(tag) => Ok(tag),
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_from_str() {
+        use crate::header::record::value::map::tag::Identity;
+
+        assert_eq!(
+            "NOODLES".parse(),
+            Ok(Other(String::from("NOODLES"), PhantomData::<Identity>))
+        );
+        assert_eq!("ID".parse::<Other<Identity>>(), Err(ParseError::Invalid));
     }
 }