+++ noodles-vcf/CHANGELOG.md
@@ -9,6 +9,8 @@
   * vcf/header/record/value/map/contig: Add getter (`Map::<Contig>::url`) and
     mutable getter (`Map::<Contig>::url_mut`) for URL.
 
+  * vcf/header/record/value/map/tag/other: Implement `FromStr` for `Other`.
+
   * vcf/record/info: Change getters to accept keys that implement
     `Equivalent<Key>`.